requirements.txt
numpy==1.26.0
pandas==2.1.1
pytz==2023.3.post1
plotly==5.16.1
dash==2.13.0
loguru==0.7.2
selenium==4.13.0
schedule==1.2.0
pyetrade==1.4.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
robin-stocks==3.0.6
tda-api==1.6.0
PyMuPDF==1.23.8